Started work on refactoring renderer and terrain

+++ Lamp/src/Platform/Vulkan/VulkanShader.h
@@ -75,7 +75,7 @@ namespace Lamp
 		inline const std::vector<ShaderDescriptorSet>& GetDescriptorSets() { return m_shaderDescriptorSets; }
 		inline const std::vector<PushConstantRange>& GetPushConstantRanges() { return m_pushConstantRanges; }
 		inline const std::unordered_map<std::string, ShaderResourceDeclaration>& GetResources() { return m_resources; }
-		inline const uint32_t GetDescriptorSetLayoutCount() { return m_descriptorSetLayouts.size(); }
+		inline const uint32_t GetDescriptorSetLayoutCount() { return static_cast<uint32_t>(m_descriptorSetLayouts.size()); }
 
 		ShaderMaterialDescriptorSet CreateDescriptorSets(uint32_t set = 0);
 		ShaderMaterialDescriptorSet CreateDescriptorSets(uint32_t set, uint32_t count);
